+++ esp-hal/src/spi/slave.rs
@@ -48,7 +48,7 @@
 //! let mut send = tx_buffer;
 //!
 //! let transfer = spi
-//!     .dma_transfer(&mut receive, &mut send)
+//!     .transfer(&mut receive, &mut send)
 //!     .unwrap();
 //!
 //! transfer.wait().unwrap();
@@ -69,9 +69,8 @@
 //! `is_done()`.
 //!
 //! See [tracking issue](https://github.com/esp-rs/esp-hal/issues/469) for more information.
-use core::marker::PhantomData;
 
-use super::{Error, FullDuplexMode, SpiMode};
+use super::{Error, SpiMode};
 use crate::{
     dma::{DescriptorChain, DmaChannelConvert, DmaEligible, PeripheralMarker, Rx, Tx},
     gpio::{InputSignal, OutputSignal, PeripheralInput, PeripheralOutput},
@@ -87,14 +86,13 @@ const MAX_DMA_SIZE: usize = 32768 - 32;
 /// SPI peripheral driver.
 ///
 /// See the [module-level documentation][self] for more details.
-pub struct Spi<'d, M, T = AnySpi> {
+pub struct Spi<'d, T = AnySpi> {
     spi: PeripheralRef<'d, T>,
     #[allow(dead_code)]
     data_mode: SpiMode,
-    _mode: PhantomData<M>,
 }
 
-impl<'d> Spi<'d, FullDuplexMode> {
+impl<'d> Spi<'d> {
     /// Constructs an SPI instance in 8bit dataframe mode.
     pub fn new<
         SCK: PeripheralInput,
@@ -108,12 +106,12 @@ impl<'d> Spi<'d, FullDuplexMode> {
         miso: impl Peripheral<P = MISO> + 'd,
         cs: impl Peripheral<P = CS> + 'd,
         mode: SpiMode,
-    ) -> Spi<'d, FullDuplexMode> {
+    ) -> Spi<'d> {
         Self::new_typed(spi, sclk, mosi, miso, cs, mode)
     }
 }
 
-impl<'d, T> Spi<'d, FullDuplexMode, T>
+impl<'d, T> Spi<'d, T>
 where
     T: Instance,
 {
@@ -130,7 +128,7 @@ where
         miso: impl Peripheral<P = MISO> + 'd,
         cs: impl Peripheral<P = CS> + 'd,
         mode: SpiMode,
-    ) -> Spi<'d, FullDuplexMode, T> {
+    ) -> Spi<'d, T> {
         crate::into_ref!(sclk, mosi, miso, cs);
 
         let this = Self::new_internal(spi, mode);
@@ -154,13 +152,12 @@ where
     pub(crate) fn new_internal(
         spi: impl Peripheral<P = impl Into<T> + 'd> + 'd,
         mode: SpiMode,
-    ) -> Spi<'d, FullDuplexMode, T> {
+    ) -> Spi<'d, T> {
         crate::into_ref!(spi);
 
         let mut spi = Spi {
             spi: spi.map_into(),
             data_mode: mode,
-            _mode: PhantomData,
         };
         spi.spi.reset_peripheral();
         spi.spi.enable_peripheral();
@@ -193,7 +190,7 @@ pub mod dma {
         Mode,
     };
 
-    impl<'d, T> Spi<'d, FullDuplexMode, T>
+    impl<'d, T> Spi<'d, T>
     where
         T: InstanceDma,
     {
@@ -316,7 +313,7 @@ pub mod dma {
         /// sent is 32736 bytes.
         ///
         /// The write is driven by the SPI master's sclk signal and cs line.
-        pub fn dma_write<'t, TXBUF>(
+        pub fn write<'t, TXBUF>(
             &'t mut self,
             words: &'t TXBUF,
         ) -> Result<DmaTransferTx<'t, Self>, Error>
@@ -351,7 +348,7 @@ pub mod dma {
         /// received is 32736 bytes.
         ///
         /// The read is driven by the SPI master's sclk signal and cs line.
-        pub fn dma_read<'t, RXBUF>(
+        pub fn read<'t, RXBUF>(
             &'t mut self,
             words: &'t mut RXBUF,
         ) -> Result<DmaTransferRx<'t, Self>, Error>
@@ -387,7 +384,7 @@ pub mod dma {
         ///
         /// The data transfer is driven by the SPI master's sclk signal and cs
         /// line.
-        pub fn dma_transfer<'t, RXBUF, TXBUF>(
+        pub fn transfer<'t, RXBUF, TXBUF>(
             &'t mut self,
             read_buffer: &'t mut RXBUF,
             words: &'t TXBUF,