+++ hil-test/tests/spi_half_duplex_write_psram.rs
@@ -14,8 +14,7 @@ use esp_hal::{
     pcnt::{channel::EdgeMode, unit::Unit, Pcnt},
     prelude::*,
     spi::{
-        master::{Address, Command, HalfDuplexReadWrite, Spi, SpiDma},
-        HalfDuplexMode,
+        master::{Address, Command, Spi, SpiDma},
         SpiDataMode,
         SpiMode,
     },
@@ -39,7 +38,7 @@ macro_rules! dma_alloc_buffer {
 }
 
 struct Context {
-    spi: SpiDma<'static, HalfDuplexMode, Blocking>,
+    spi: SpiDma<'static, Blocking>,
     pcnt_unit: Unit<'static, 0>,
     pcnt_source: InputSignal,
 }
@@ -65,7 +64,7 @@ mod tests {
 
         let mosi_loopback = mosi.peripheral_input();
 
-        let spi = Spi::new_half_duplex(peripherals.SPI2, 100.kHz(), SpiMode::Mode0)
+        let spi = Spi::new(peripherals.SPI2, 100.kHz(), SpiMode::Mode0)
             .with_sck(sclk)
             .with_mosi(mosi)
             .with_dma(dma_channel.configure(false, DmaPriority::Priority0));
@@ -99,7 +98,7 @@ mod tests {
         // Fill the buffer where each byte has 3 pos edges.
         dma_tx_buf.fill(&[0b0110_1010; DMA_BUFFER_SIZE]);
         let transfer = spi
-            .write(
+            .half_duplex_write(
                 SpiDataMode::Single,
                 Command::None,
                 Address::None,
@@ -113,7 +112,7 @@ mod tests {
         assert_eq!(unit.get_value(), (3 * DMA_BUFFER_SIZE) as _);
 
         let transfer = spi
-            .write(
+            .half_duplex_write(
                 SpiDataMode::Single,
                 Command::None,
                 Address::None,
@@ -151,7 +150,7 @@ mod tests {
 
         let buffer = [0b0110_1010; DMA_BUFFER_SIZE];
         // Write the buffer where each byte has 3 pos edges.
-        spi.write(
+        spi.half_duplex_write(
             SpiDataMode::Single,
             Command::None,
             Address::None,
@@ -162,7 +161,7 @@ mod tests {
 
         assert_eq!(unit.get_value(), (3 * DMA_BUFFER_SIZE) as _);
 
-        spi.write(
+        spi.half_duplex_write(
             SpiDataMode::Single,
             Command::None,
             Address::None,